Fix 32-bit ISE 10.1 on 64-bit platform.
When trying to access a cable with this combination, ISE displays the
following error: "Cable operation is not supported when running the
32-bit version of the application on a 64-bit platform."

This is fixed by additionally overloading the uname function and replacing
x86_64 by i686 in the returned struct in the 32-bit version of this library.
The 64-bit version will still correctly return x86_64.